xiph demux: Kill a warning (comparison betw. signed and unsigned)

+++ modules/demux/xiph.h
@@ -48,7 +48,7 @@ static inline int xiph_SplitHeaders(unsigned packet_size[], void *packet[], unsi
         }
         size += packet_size[i];
     }
-    if (end - current < size)
+    if (end - current < (int)size)
         return VLC_EGENERIC;
     packet_size[count - 1] = end - current - size;
